backend/requirements.txt
# Smart Parking Platform v6 - Backend Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23  # Optional - for ORM (currently using raw SQL)

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis (for downlink queue and caching)
redis==5.0.1
aioredis==2.0.1

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1

# ChirpStack Integration
grpcio==1.60.0
protobuf==4.25.1
# chirpstack-api==4.5.1  # Uncomment when needed

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
